acw-ui/requirements.txt
altair==5.1.1
attrs==23.1.0
backports.zoneinfo;python_version<"3.9"
blinker==1.6.2
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
cloudevents==1.9.0
cryptography==41.0.3
deprecation==2.1.0
firebase==4.0.1
firebase-admin==6.2.0
Flask==2.3.3
functions-framework==3.4.0
gitdb==4.0.10
GitPython==3.1.34
google-api-core
google-api-python-client==2.99.0
google-auth==2.22.0
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos
grpcio==1.58.0
grpcio-status
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
msgpack==1.0.5
numpy==1.24.4
packaging==23.1
pandas==2.0.3
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
proto-plus==1.22.3
protobuf
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydeck==0.8.0
Pygments==2.16.1
PyJWT==2.8.0
Pympler==1.0.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-firebase==1.2
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
referencing==0.30.2
requests==2.31.0
rich==13.5.2
rpds-py==0.10.2
rsa==4.9
semver==3.0.1
six==1.16.0
smmap==5.0.0
streamlit==1.26.0
streamlit-custom-notification-box==0.1.2
tenacity==8.2.3
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
uritemplate==4.1.1
urllib3==1.26.16
validators==0.22.0
watchdog==3.0.0
Werkzeug==2.3.7
zipp==3.16.2
plotly_express